Prints the last digit of a number

+++ 0x02-functions_nested_loops/7-print_last_digit.c
@@ -9,7 +9,7 @@
  */
 int print_last_digit(int x)
 {
-	unsigned int lastDigit;
+	long lastDigit;
 
 	if (x < 0)
 	{
